Onboard CFF2
This onboards the in-development CFF2 to the monorepo. We are going to
hold off on writing BUCK macros for use in the monorepo until this is
ready for consumption by others.

I had to disable the golden_test because the monorepo indiscriminately
reformats checked in Go files, which changes the expectations for the
golden test. We can fix this after this lands.

Reviewed-on: D2551737