example/magic: Use InstrumentFlow
The prior examples/magic.go file didn't compile because it was
attempting to use cff.Instrument instead of InstrumentFlow to instrument
the task flow.

We *can* introduce a new Option type that applies to both, tasks and
flows, at which point you would be able to use Instrument for both, but
that will likely hurt discoverability in godocs.

Reviewed-on: D2553861