gen: Fix shadowing global err variable
The generated code for multiple concurrent tasks were both concurrently
modifying the global 'err' variable. Moreover, there was some existing
code for creating a per-task error variable but we were only reading,
never writing to the variable, and the variable was shared between the
scheduled task group rather than a per-task variable. Fixes the
generated code to use per-task error variables and when an error
occurs copy from the per-task variable to the global error variable.

Reviewed-on: D2570951