Remove special case for n=1 tasks
Panic handling necessitates using defer(), however if we have a special
case for n=1 tasks where we don't use separate goroutines it would
complicate the recovery logic to infer *which* task was running when the
panic occurred. There would be two choices without this diff, both have
bad trade-offs. As a nice side effect, this reduces the code duplication
in the template so that there aren't two places to update when you
change what happens before, during, and after a user task runs.

Either we would have one defer() per group of tasks, and leverage the
fact that defers are run in reverse order so that the most recent defer
would be the one that was caught. Because it's not possible to "undo"
a deferred function, there would be a number of deferred recovery
functions that are essentially noops. This is wasteful since we really
only should be recover()'ing once.

The other choice is to have a single defer() with some mutable variable
like "lastRunTask". While this doesn't incur any waste, it is slightly
more complicated than this change.

Reviewed-on: D2602121