compile: errf to require source code AST node position
Small refactor split out from error handling diff D2627443. We will want to separate out the presentation of the error (including the line and absolute file path on which it occurred) from the error information itself, so change the signature of `errf` to take the node position and centralize the logic for formatting that into `errf` so it can be easily moved in the future to specific error types.

Reviewed-on: D2631121