Add test case for nested taskflows
This mostly works as expected, however there is a small IDE problem
with go-to-definition. When you're in the parent's flow definition, going to
definition of nested_child.Itoa goes to the definition of the *generated*
child.Itoa rather than the CFF source. This may be a quirk of Goland since
technically the generated code has a build tag !cff. Added a TODO to track this
minor issue.

Reviewed-on: D2628881