compile: Fix errors accumulating multiple times
The previous diff D2649009 didn't account for the compiler also
accumulating errors in c.errors, so when an error occurred in file A and
B, we would accumulate A and B in c.errors but accumulate A twice in
Process. Instead we trust compiler to accumulate errors and pull them
out at the end in Process.

Reviewed-on: D2649283