More bad input tests
Adds back the logic to cff.Flow because while the Go compiler will
verify that it's a cff.FlowOption, there's nothing stopping the user
from doing a few tricks to get a value that is a cff.FlowOption but not
how we want it. Adds a few more test cases to get the coverage up.

Reviewed-on: D2675127