Use go-flags
This changes CFF to use the go-flags command line parser instead of the
standard library's flag package.

This is a more versatile command line parser and will make some future
changes cleaner.

While we're at it, I also changed the Makefile to use bazel instead of
buck to run its commands. (I can move this to its own diff but it didn't
seem worth it in this case.)

Reviewed-on: D2920185