Refactor of CFF rule to generate an output file
Refactoring CFF rule to treat CFF source as in-place replaceable and
output the transpiled CFF file instead of a library like before. This
will allow other code generation systems to use this rule on top of
theirs.

This change also allows support for multiple CFF flows and functions
within the same package. Adding a test for that case as well.

CFF CLI is also changing with new flags to require a specific file to be
inspected along with entire package containing CFF sources.

More details for test case we are covering in T3059641.

Reviewed-on: D2884309