Rename flow and task metric tags to "flow", "task"
Fixes T3167041

"name" is a M3 reserved word, and metrics that are emitted to the M3
aggregator will not be correctly processed, so some of the stats are
dropped. Rename the metric tag keys to non-reserved words, "task" for
task metrics and "flow" for flow metrics respectively. Also change
log keys for uniformity.

Rename `failedTask` tag to `failedtask` since M3 will lower case it anyway.

Reviewed-on: D2936523