Improve and annotate unit tests for cff.Task(s)
Currently it's hard to tell which error message corresponds to which tests. After classification I discovered that some unit tests can be commented out and won't trigger errors since those errors are matched in a different Flow. Fixing that, and also noting that one case is also not triggered for similar reasons.

Reviewed-on: D3131251