Fix no-output tasks not handling FallbackWith annotation.
Fixes T3618933

The template uses the variable .FallbackWith to check if we need to
generate the code related to suppressing the error of the task. For
no-output tasks, the .FallbackWith variable will be a list of zero
length which is falsy ("The empty values are ... any slice ... of
length zero"). Change it so that we can track whether we need to
suppress the error for FallbackWith independent of the list of
expressions that provide the value for each return type of the
function, so that no-output tasks can support FallbackWith; no-ouput
tasks will by definition always have no return types so
FallbackWithResults must be empty.

Reviewed-on: D3142309