Fix flowtags having failedtask tag when using RecoverWith
Fixes T3630161. Due to some code duplication, we mistakenly mutate
flowtags which are the tags that are used in the final metric for the
taskflow when we are recovering an error with RecoverWith. This
results in flows that use RecoverWith having many metrics like
flow.success+failedtask=XYZ. Remove the flowtags mutation so we just
emit flow.success.

Reviewed-on: D3150007