Automatically infer a placeholder name for a non-instrumented task
Fixes T3702633

We currently have a number of flows that are used in production but don't have instrumentation enabled for some of the tasks.

The requirement to define names for instrumented tasks is sometimes tedious, and if you do not do this you miss out on:
- Stack traces for panics
- Logs (specifically of errors and panics)
- Metrics (specifically for errors and panics)

This can complicate debugging things like why a flow is erroring or panicking (e.g. T3609667)

We are missing a linter (T3304557) that enforces that Instrument is used as well, so it can slip through code review and we end up missing instrumentation when we need it.

As a **stop-gap** for the better solution which will come when we (1) rewrite the observability interface to support user-provided instrumentation hooks, and (2) have the linter (T3304557), we can infer a basic name for the tasks that are missing instrumentation names, with the obvious caveats:

1. The implicit names will change when the file is refactored because it uses line numbers,
2. The implicit names are not necessarily unique because two files may have the same name,
3. The implicit names are also not unique because the user could define a flow that is on a single line (unlikely), and
4. There will be extra metrics for tasks that "don't need instrumentation" although I would argue this point.

(2) and (3) happen to be not that a big deal for us because we add tags to the metrics so at least they will be scoped by procedure name and flow name.

The default inferred placeholder name would be something like the following:

```
// test.go

1:
2: func Foo() {
3:     cff.Flow(context.Background(),
4:         cff.Logger(...),
5:         cff.Metrics(...),
6:         cff.Task(func() { }))
7: }
```

The task will be instrumented as if the user had written:

```
// test.go

1:
2: func Foo() {
3:     cff.Flow(context.Background(),
4:         cff.Logger(...),
5:         cff.Metrics(...),
6:         cff.Task(func() { }, cff.Instrument("test.go.6")))
7: }
```

For examples of flows that are missing instrumentation on some of the tasks:

https://code.uberinternal.com/diffusion/GOCODVJ/browse/master/src/code.uber.internal/presentation/apps/rider/helix/endpoints/status/get-status/handler.go$98-100

https://code.uberinternal.com/diffusion/GOCODVJ/browse/master/src/code.uber.internal/presentation/apps/rider/tasks/status/postprocessor/task.go$91

https://code.uberinternal.com/diffusion/GOCODVJ/browse/master/src/code.uber.internal/presentation/apps/rider/helix/endpoints/surge/batch/handler.go$105

Reviewed-on: D3142879