Fix implicit types for task dependencies or outputs not importing needed package
Fixes T3989113

When a user defines a flow using a non-literal function like
cff.Task(foo) instead of cff.Task(func(baz.Foo) ...), the generated
code that will be produced will have references to the types needed
for foo in the variables (e.g. var v1 = baz.Foo). But the user may not
have imported baz. This change checks when we are forming the
qualifier for those variables whether the package is already imported
and if not adds it to the list of added imports so that the qualifier
is not referring to an unimported package.

Reviewed-on: D3694017