Add MetricsEmitter interface for providing custom metrics observer
Adds a new flow option and API for users to provide custom code for observing flow events. The default implementation is limited to metrics and it memoizes the scope based on the task and flow names.

We evaluated three approaches around TaskFlowFailed: a `Mutex`, a `sync.Once`, and [double-checked locking](https://en.wikipedia.org/wiki/Double-checked_locking) where we set a read lock for contending tasks using shared state which returns if a task has already failed. However, benchmark results below showed that a single `Mutex` provides better performance until we create benchmarks to best model the contention case over the write lock. An example is a task that fails early running in parallel with ten tasks that will eventually fail at the same time (e.g. timeout); the latter ten tasks will contend on the Mutex while a double-checked lock would allow all of them to hold a RLock to check that the failed task flag is set (using `Failed()`), while with the current implementation all of them will serially check `Failed()`. No significant difference was found between `sync.Once` and `Mutex`, so we switched to `sync.Once` for more idiomatic code.

Test Plan:
Benchmarks show that the diff is no worse than master except for a slight increase in allocations when not using the new optional APIs when number of tasks is small.

|**Benchmark**|**ns/op**|**ns/op %**|**B/op**|**B/op %**|**allocs/op**|**allocs/op %**|
|Baseline-8|-143|-8.502|0|0.000|0|0.000|
|Metrics/Metrics-8|-1102|-5.726|-1416|-23.077|+3|+4.286|
|Metrics/Metrics100-8|-101049|-30.489|-59456|-36.641|-284|-17.349|
|Metrics/Metrics500-8|-277696|-19.657|-275045|-34.419|-1471|-18.303|
|Metrics/Metrics1000-8|-279592|-11.360|-549342|-34.470|-2949|-18.407|
|Metrics/Metrics1000Failed-8|-935901|-32.916|-549867|-35.170|-2955|-17.343|

When users opt in to the new APIs, compared to `Metrics1000` and `Metrics1000Failed` respectively:

| **Benchmark** | **Δ ns/op** | **Δ ns/op %** | **Δ bytes/op** | **Δ bytes/op %** | **Δ allocs/op** | **Δ allocs/op %** |
|MetricsMemoized1000-4       |	 -1411621 | -62.51% | -1472353 | -93.39% |  -14002 | -87.40% |
|Metrics1000FailedMemoized-4 |	-1586037 | -75.98% | -1472102 | -94.20% | -14001 | -82.21% |

Raw results: P207619 for master, P207621 for this diff.

Refs T4783413
Reviewed-on: D3884115