Add WithLogFields option
This adds a WithLogFields flow option to provide per-request log fields
for CFF flows.

For example,

    cff.Flow(..., cff.Logger(log), cff.WithLogFields(a, b, c))

Will yield behavior similar to,

    cff.Flow(..., cff.Logger(log.With(a, b, c)))

This option is needed so that as we move logging to the emitter API
(D3924301) users will need to be able to supply per-request log fields.
That is, users will be able to turn the following,

    cff.Flow(..., cff.Logger(log.With(a, b, c)))

Into,

    cff.Flow(...,
        cff.WithEmitter(emitter),
        cff.WithLogFields(a, b, c))

---

This has no effect on performance. Differences in runtime are within the
range of expected variance.

|**Benchmark**|**ns/op**|**ns/op %**|**B/op**|**B/op %**|**allocs/op**|**allocs/op %**|
|Baseline-4|-7|-0.452|0|0.000|0|0.000|
|BaselineNative-4|-7|-0.619|0|0.000|0|0.000|
|Metrics/Metrics-4|-84|-0.464|0|0.000|0|0.000|
|Metrics/Metrics100-4|-2251|-0.771|+4|+0.004|0|0.000|
|Metrics/Metrics500-4|+7382|+0.545|-22|-0.004|-1|-0.015|
|Metrics/Metrics1000-4|-90488|-2.873|-41|-0.004|0|0.000|
|Metrics/Metrics1000Failed-4|-373117|-14.001|-603|-0.059|-6|-0.043|
|Metrics/MetricsMemoized1000-4|+50257|+3.430|0|0.000|0|0.000|
|Metrics/Metrics1000FailedMemoized-4|-29891|-4.006|+157|+0.173|+1|+0.033|

Reviewed-on: D3927785