gen: Simplify Emitter creation
We currently generate code in the following form when an Emitter is not
provided.

    func(ctx context.Context, scope tally.Scope, logger *zap.Logger) (err error) {
        var emitter cff.Emitter
        emitter = cff.DefaultEmitter(scope)
        ...
    }(ctx, scope, logger)

And if one is provided, we generate,

    func(ctx context.Context, scope tally.Scope, logger *zap.Logger) (err error) {
        var emitter cff.Emitter
        emitter = userProvidedEmitter
        ...
    }(ctx, scope, logger)

Note that we ignore the `scope` argument.

This changes the generated function to always accept a `cff.Emitter` if
observability is enabled so that generated code for the two cases looks
like the following.

    func(ctx context.Context, emitter cff.Emitter, logger *zap.Logger) (err error) {
        ...
    }(ctx, cff.DefaultEmitter(scope), logger)

    func(ctx context.Context, emitter cff.Emitter, logger *zap.Logger) (err error) {
        ...
    }(ctx, emitter, logger)

This has the effect of simplifying the template and producing cleaner
output.

Reviewed-on: D4147143