Disallow mixing WithEmitter and Metrics
If a user uses both, `cff.WithEmitter` and `cff.Metrics`, we ignore
`cff.Metrics` and use the `Emitter`. Rather than ignoring it silently,
we should generate an error informing the user of this because this was
likely a mistake.

This changes instrumentation validation to fail if `cff.Metrics` was
provided with `cff.WithEmitter` and updates test cases that were
unknowingly doing so.

Reviewed-on: D4149421