Take a pass over the docs
This takes a full pass over the CFF2 docs, rewriting and restructuring
as necessary.

The flow of information is like so:

- The home page lays down the basic concepts and terminology.
- Getting Started demonstrates a bare minimum usage of CFF2 without
  going into all the different things you can do with tasks. This page
  covers nothing but the bare minimum features.
- The Tasks page details the various different things tasks can do,
  including the flexbility offered in signatures of task functions.
- Observability, authored in D4160221, continues to document metrics and
  logging for CFF2. Note that this does not yet document emitters
  because we're not done messing with the Emitter APIs.

This is obviously not perfect and there's room for expansion in other
places. We can, for example, add a separate page about the `cff` Bazel
rule and its various parameters. That said, since this Diff is a net
positive on top of what we have, I recommend we defer new sections to
future diffs.

Changes to the API Reference:

- For every annotation (now referred to as a *code generation
  directive*) we now mention that the `cff` build tag is necessary for
  files using that function. This was to differentiate APIs in the cff
  package that can be used from any Go file (`cff.Emitter`, for
  example).
- The documentation for `Fallbackwith` was inaccurate. It stated that
  CFF2 does not use fallback values for `panic`s, but the code says
  otherwise.

Test Plan:
=== Home ===

{F250770059}

=== Getting Started ===

{F250770557}

=== Tasks ===

{F250769901}

Reviewed-on: D4163511