emitter: Split TaskRecovered for Error, Panic
=== CFF ===

This change breaks TaskRecovered into two events: TaskErrorRecovered and
TaskPanicRecovered. This is necessary because we need to differentiate
between and log different messages for "task errored," "task panicked,"
"task errored and recovered," and "taks panicked and recovered."

This wouldn't be possible with the current method of doing the
following.

    e.TaskError(ctx, err)
    e.TaskRecovered(ctx, err)

The emitter has no way of knowing whether TaskError should log
or TaskErrorRecovered should log.

With this change, the above becomes,

    e.TaskErrorRecovered(ctx, err)

And similarly, recvovered panics become,

    e.TaskPanicRecovered(ctx, err)

Something to consider in the future: Whether the failure was recovered
or not can be a parameter on TaskError/TaskPanic.

=== Presentation ===

Updated generated code.

Reviewed-on: D4170067