emitter: Move Default into TallyEmitter
This moves the implementation of the default emitter into TallyEmitter
almost verbatim. The idea is that we will implement a separate
LogEmitter (and possibly a JaegerEmitter) that users can pick and choose
from.

The DefaultEmitter constructor will compose these to provide the most
common configuration, which we can possibly export from an Fx module in
the future.

Reviewed-on: D4176491