Implement Log Emitter
This imlements an Emitter that logs messages to a Zap logger. This
emitter is not used anywhere at this time, but we will be able to switch
users over to it in the future.

The behavior of this emitter is exactly the same as what the generated
code does right now with one exception: we include names of the flows in
task-level logs. This aligns with a prior change to tag metrics with
names of the flows.

We test log parity by ensuring that logs emitted by the LogEmitter are
the same as CFF2's generated code, optionally with newer information.

Reviewed-on: D4181811