LogEmitter: Align with presentation logs
This change:

- adds support to the CFF2 LogEmitter for customizing the levels used to
  log errors, panics, and recovered failures
- aligns log messages and fields with presentation's emitter

The two changes together allow deletion of the custom emitter in favor
of the standard LogEmitter. This allows us to lock the interface down
again while we iterate on it.

Users of the LogEmitter will see the following differences from CFF2's
current behavior:

- "taskflow" now referred to as "flow" in logs
- Task/FlowDone entries are logged now at debug level
- FlowError now logs the name of the failed task
- panics log `panic-value` instead of coercing to errors
- unrecoverable errors are logged at debug level

Presentation will see the following differences:

- panics--with or without FallbackWith--will include a stack trace
- panics will not be coerced into errors, and instead will always be
  logged as `panic-value`

Reviewed-on: D4181813