Remove unused WithLogFields
WithLogFields was intended to allow plumbing per-request log fields at
the cff.Flow callsite. In hindsight, this introduces a design smell that
the Emitter interface, which is agnostic of what it's used for would now
know about Zap fields.

We can revert this change because it's unused and users have two other
methods to accomplish the same:

- attach fields to the logger when the LogEmitter is constructed
- use ZapFx's ability to extract fields from context, and plumb the
  fields via context

Reviewed-on: D4183003