compile: Simplify invalid directive detection
We currently maintain a whitelist of functions defined in the cff
package that can be called at the top-level. This was the preferred
approach when CFF2 had no runtime APIs, only compile-time.

As we're adding a number of runtime APIs, it is no longer desirable to
maintain a whitelist of functions that may be called directly.

This change switches to maintaining a blacklist of code generation
directives that may not be called at the top-level inside a CFF-tagged
file.

Reviewed-on: D4182411