Generate list of directives automatically
Due to a recent change, we had to manually manage the list of code
generation directives available with CFF2 in the "directives.go" file.

This change deletes that in favor of an equivalent file generated
automatically by inspecting cff.go.

Going forward, as long as all code generation directives are added
directly to the cff.go file, the list of directives will always be
up-to-date.

Reviewed-on: D4209817