Delete FlowFailedTask function part of Emitter interface.
The failedtask tag that is emitted as part of taskflow.error counter metric is now deprecated and will no longer be emitted on 2020/04/13. The same information can be found in the new flow tag emitted for task.error metric. This allows you to continue to find, for a particular flow, which task caused it to fail. Users are recommended to update their dashboards, as described below.

For example, current M3QL query

    fetch service:foo name:taskflow.error | sum flow failedtask

    | alias {{.flow}} {{.failedtask}}

Could be rewritten as the following:

    fetch service:foo name:task.error | sum flow task

    | alias {{.flow}} {{.task}}

This change is being done to simplify the implementation of CFF2 to be more performant.

Refs T5490543
Reviewed-on: D4208779