compile: Single place to construct instrument
The intent behind the `compile*` methods is that they're the
constructors responsible for building and validating their corresponding
objects. So, `compileInstrument` is the only place where an `instrument`
object should be constructed.

That way, if there is need to validate all instrument invocations
somehow, that can be in one place rather than tracking down every
`instrument` invocation.

This changes the compiler so that instrument is no longer built outside
the `compileInstrument*` functions.

Reviewed-on: D4229991