compile: Initialize flow in compileFlow only
We should avoid calling methods on an object from inside its
constructor. As setting up emitters is part of the compileFlow logic, it
doesn't need a separate method.

This drops the addEmitters method in favor of setting that inside
compileFlow.

In the future, if the logic for what constitutes an emitter becomes more
complicated, we can write a constructor for it similar to
compileInstrument.

    e := c.compileEmitter(ce.Args[0])
    flow.Emitters = append(flow.Emitters, e)

For now, grabbing the `ast.Expr` as-is suffices.

Reviewed-on: D4229993