gen: Assume non-nil emitters always
This changes the CFF template to set up nop flow and task emitters when
flows or tasks are instrumented.

This allows for the rest of the template to be simplified because it can
assume that the emitters are always non-nil instead of checking that a
task/flow is instrumented before adding the event emission block.

Refs T5698923
Reviewed-on: D4229995