gen: Declare tasks in topological order
CFF declares the task in the order they were specified before scheduling
them.

Given the following,

    cff.Task(f1),
    cff.Task(f2),
    cff.Task(f3),

The generated code takes the following shape:

    t1 := func(..) { f1() }
    t2 := func(..) { f2() }
    t3 := func(..) { f3() }

This introduces complexity in the generated code if t1 depends on a
result produced by t3 because the variable must be declared in advance.
That is, because the generated code will take the following form,

    t1 := func(..) { v1 = f1(v3) }
    // ...
    t3 := func(..) { v3 = f3()   }

v3 must be declared before t1.

To work around this, CFF declares all output variables separately before
declaring individual tasks.

    var (
        v1 T1
        v2 T2
        v3 T3
    )

    t1 := func(..) { v1 = f1(v3) }
    t2 := func(..) { v2 = f2(v1) }
    t3 := func(..) { v3 = f3()   }

This commit changes CFF to calculate a topological ordering of the tasks
at compile time so that tasks are declared before tasks that consume
them. (In this order, tasks can conceivably be run serially.)

    var v3 T3
    t3 := func(..) { v3 = f3()   }

    var v1 T1
    t1 := func(..) { v1 = f1(v3) }

    var v2 T2
    t2 := func(..) { v2 = f2(v1) }

Advantages:

- The task template has to make fewer assumptions about what's available
  in the flow scope.
- Scope of variables is reduced: we put off declaring them until they're
  needed.
- We'll need access to a topological ordering of tasks in the future
  anyway for the runtime scheduler.

Worth noting that VGF tasks are already in topo-sorted order, presumably
because that's how VGF generates the CFF code.

Reviewed-on: D4231069