test: Make BUILD.bazel Gazelle-friendly
The current cff/BUILD.bazel is not very Gazelle friendly. Adding new
tests to the root package is difficult because Gazelle will not pick up
these new tests fields and their dependencies automatically.

This reorganizes it a bit by putting the standard "package cff" tests
under the `go_default_test` rule--this is the rule Gazelle manages
automatically.

instrument_test is currently the only test that needs to be in its own
package because it relies on CFF-generated code, which relies back on
the CFF package. So we place that in its own test rule with a name that
is //not// automtatically managed by Gazelle. The `gazelle:exclude`
clause makes sure that Gazelle does not add it to the `go_default_test`
rule.

Testing: Ran gazelle. Notice the absence of `# keep` comments.

Reviewed-on: D4234999