gen: Pull static task executor into library
As suggested in D4230637, this pulls the logic to execute a set of
statically scheduled tasks into a CFF library level function. Besides
reducing code coverage for the generated code, this allows us to test
the core execution logic more exhaustively and separately from the
generated code.

Note that this does introduce a small behavioral change: previously, a
cancelled context would not result in a FlowError event, only
FlowSkipped. With this change, it will emit FlowError too.

This should be fine because the FlowSkipped metric is redundant. It
almost always has the same value as FlowError. We can probably delete it
in the future.

Refs T5698923

Reviewed-on: D4231865