Drop bindata/ prefix
The bindata/ prefix was largely a hack to help Go modules work. Now that
Go modules are used only for dependency resolution, bindata prefix isn't
necessary.

Change bindata-generated package to be at the same path as the actual
templates so that all CFF code is under
go.uber.org/cff.

Reviewed-on: D4415039