compiler: Hide complierOpts, don't embed
The pattern for constructors is `func New(options) thing` where `New`
pulls relevant informatiion from `options`. Putting `options` in `Thing`
wholesale adds an unnecessary level of nesting.

Align with this by switching `newCompiler` to accept just `compilerOpts`
and pulling information from it onto `compiler`.

Further, the compiler abstraction is not known to the users of the
internal package (compiler is not exported), so compilerOpts makes
little sense as an exported type.

Resolve this by unexporting compilerOpts and instead, exporting a new
Processor on which these options may be configured.

Reviewed-on: D4415027