Implement runtime scheduler
This implements a runtime job scheduler for CFF with the following
properties:

- Variable number of workers, decoupled from the number of jobs
- Lockless scheduler loop
- Jobs may defer their execution until other previously scheduled jobs
  have finished
- Failure in any job halts execution

This does not integrate the scheduler into CFF yet. Once integrated, the
generated code will use the scheduler roughly in the following form:

    sched := scheduler.Begin(..)

    j1 := sched.Enqueue(ctx, scheduler.Job{
        // ...
    })

    j2 := sched.Enqueue(ctx, scheduler.Job{
        // ...
        Dependencies: []*scheduler.ScheduledJob{
            j1,
        },
    })

    err := sched.Wait(ctx) // wait for all enqueued jobs to finish

Jobs declare dependencies on previously enqueued jobs with the
Dependencies field. This keeps the responsibility of dependency
resolution on the caller, adding no runtime cost for the scheduler.

The scheduler works by keeping track of jobs with no outstanding
dependencies (these are ready to run), and jobs with outstanding
dependencies (waiting to run). Whenever a job finishes running, all jobs
waiting for it are notified, and those for which this was the last
dependency are moved to the ready-to-run queue.

The visibility of the scheduler package is limited to just CFF and its
subpackages to prevent misuse of these APIs.

Refs T4044393

Reviewed-on: D4405931