Add opt-in for online scheduler
Add an `--online_scheduling` flag and an `online_scheduling` Bazel rule
flag to opt into the CFF2 online scheduler.

To avoid requiring new imports in the BUILD.bazel files for all CFF2
users, re-export relevant APIs from the top-level `cff` package.

---

With the new scheduler, it's possible that a worker is still running a
task when another task errors out. When a task fails, `sched.Wait` will
return early. This introduces a race for the other running task: it may
attempt to set `task.ran` to true at the same time that the flow is
reading `task.ran` to see if `TaskSkipped` needs to be called.

     |      sched.Wait      task1       task2
     t          .           |           |
     i          .           |           |
     m          : <--------<! error     |
     e          |                       |
     |          |              race     |
     v      if !task2.ran  - - - - - - -+ task2.ran = true
              e.TaskSkipped

Resolve this race by switching `task.ran` to an atomic boolean.

Refs T4044393

Reviewed-on: D4405933