Fix latency metrics emitted when predicate evaluates to false
Fixes T6278905. We unconditionally call taskEmitter.TaskDone when a
task completes. Task latency metrics are emitted as part of this
call. When a predicate is false, we will also emit task latency
metrics, which conflict with the task latency metrics when predicate
is true. This commit changes it so we just don't call TaskDone when
predicate is false, so no latency metrics are emitted when a predicate
is false.

There will still be task.skipped logs and metrics when predicate
evaluates to false. This also changes log emission when predicate is
false: instead of emitting both "task.done" and "task.skipped", we
simply emit "task.skipped".

Coverage failing due to GM-75

Reviewed-on: D4649739