renable CFF test
When there is an unexpected call to a gomock in a non-main go routine, gomock
kills the go routine which means the code to unblock the test does not run
which hangs the test (https://github.com/golang/mock/issues/145).

'TestPanic' in cff/instrument_test.go hangs because TaskDone is unexpectedly
called in a non-main go routine.

Even if the task panics, that task should be considered done and not skipped
(GO-258). This updates the test expectations accordingly and addresses
the unexpected call.

ref: GO-303

Reviewed-on: D5307809