Introduce a function to the dependency graph
The current CFF dependency graph is centered around the 'Task' object.
However, the Task object is not flexible enough easily represent
abstractions like cff.Predicates as part of the dependency graph.

This diff introduces the concept of a function as a smaller, more generic
unit of abstraction and execution in the CFF dependency graph. With this new
abstraction, it becomes easier to introduce new objects to the graph such as
cff.Predicates.

Note: This change does not modify any existing exposed functionality or
generated code.

resolve GO-260

Reviewed-on: D5495157