Separate predicate execution from task
Separate out the execution of a predicate from the task that the predicate is
a condition for in the CFF dependency graph.

This is a performance optimization as previously a predicate was executed
within a task and therefore blocked on that task's execution. After this
change, predicates can now be executed before and outside of the task they are
a condition for and take advantage of idle workers they previously could not
be scheduled for.

resolve GO-260

Reviewed-on: D5495165