predicate: fix slow dep test flakiness
While running the scheduler test cases with --runs_per_test=1000 and
--cache_test_results=no I found that 'TestPredicateScheduling' was
flakey as the defer statement to assert that slowDepDone.Load() was
false ocassional failed.

Asserting slowDepDone.Load() as false tests that the predicate job
(that does not depend on on the slowDep job) is able to run before the
slowDep job and therefore run more optimally. It's important to note that the
slowDep jobs is a dependency of the job that the predicate job is a
condition for.

To address the race condition, within the predicate run function, I unblock the
slowDep job after the assertion that the slowDep job has not finished instead
of the regular function body so that the predicate is certainly done before the
slowDep job finishes.

Refs GO-686
Reviewed-on: D6259241