Update test importer's exporter
In the test importer, we have something called an exporter, which helps
us write testdata into a workspace. We were previously using the
packagestest.Modules exporter, which puts each module into its own
repository, whereas we want packagestest.GOPATH exporter, so that it
emulates a GOPATH.

This broke for Glue in the Go1.16 upgrade, but did not break for CFF
as CFF only exports a single module. To prevent future confusion, this
aligns the exporter setting with what we have in Glue (D6438231).

While we are here, clean up the documentation of the compile test setup
function.

Refs GO-84
Reviewed-on: D6446335