parallel prefactor: FlowOption to Option
Rename FlowOption to Option

cff.Flow is configured with a FlowOption type. When cff.Parallel is introduced,
a natural path is to configure cff.Parallel with a ParallelOption type.

Given that cff.Parallel and cff.Flow will share some configuration options,
this action shares code by unifying the options for cff.Flow and cff.Parallel
under a single Option type.

Refs GO-84
Reviewed-on: D6446345