Use code.uber.internal/go/importer package for CFF2
This removes the internal importer implementation CFF in lieu of
the code.uber.internal/go/importer package.

It is mainly a 1-to-1 replacement of the corresponding APIs.

Following up to this diff will be some additional cleanup of the
go/importer package API surface, as well as removing some remnant
package related logic in CFF into the go/importer package, such as
(but not limited to):

- Parsing a rules_go convention Archive path string.
- Wrapper for loading types.Package using x/tools/go/packages.

Refs GO-589
Reviewed-on: D6569361