Remove codegen logic to flip build tag
Following the closure of GO-147, build tags for cff are no longer
required to ensure non-generated cff code from successfully
(and incorrectly) building and deploying. As a consequence, this
commit removes the logic in `marketplace/cff` code generator
that flips the build tag in the generated code. This logic is no longer
needed.

In addition to this, removal of this build tag helps resolve
build errors in certain targets when upgrading to go1.17
in the monorepo due to multiple conflicting build tags.
See GO-903 for details.

Refs: GO-196, GO-903

Reviewed-on: D6648763