Make SchedulerInfo directive agnostic
SchedulerInfo in its current form embeds a FlowInfo. To emit scheduler metrics
for cff.Parallel, a natural path would be to also embed a ParallelInfo to the
SchedulerInfo struct. However, this complicates the code as each of the
individual emitter implementations (e.g log emitter and tally emitter)
would need to add internal logic to switch paths based on the type of
directive (e.g. flow or parallel).

Given that observability at the level of Scheduler is at an abstraction
below information unique to a cff.Flow or cff.Parallel, I flatten the
SchedulerInfo struct with a generic "directive" concept that statisfies a
flow, parallel, or furture directives. Additionally, i've modified the
emitter implementation to work with this more generic set of information.

Refs GO-84
Reviewed-on: D6647881