aquaregia fix: Extract out test cases
The aquaregia test, which ensures that CFF compile errors are surfaced
correctly, contain a list of errors to expect from the compilation
process. In preparation for adding a test which checks if the CFF DSL
compiles correctly, separate out these cases into a separate function so
they can be re-used by other tests.

Refs GO-954
Reviewed-on: D6722143