aquaregia feature: Test error match frequency
The aquaregia test, which ensures that CFF compile errors are surfaced
correctly, only asserts whether an error has appeared in the
specific file and does not verify the number of times an error has
occured relative to it's declared expected frequency.

This gap in testing has led to passing test expectations about the
frequency of observed errors that are not observed in test behaviour.

This diff fixes all outstanding error match frequency expectation issues and
adds logic to check if the frequency of expected errors is seen in the errors
returned by the CFF compiler.

Refs GO-950
Reviewed-on: D6722147